Moving the mediatorUrl, mediaPlayerSettings to lib directory & moving logic for generating AV settings out of ArticleMain into same directory
